Better handle payment completion.
I have no idea how this was *supposed* to work before, since customers are returned from Fetch without any data in `$_POST['txn_success']` ... but what it does now is compare the order status to what the IPN would set the order status to.

The previous implementation just plain didn't work though so this is a bit of a leap forward :)

This may not be as robust as it could be, hit me up if you'd like to improve it further.